Cambio de nombres de clases y finalización del carrito de compras
- Se renombró la clase 'Sale' a 'Order'.
- Se renombró la clase 'SaleDetail' a 'OrderDetail'.
- Se completó la funcionalidad del carrito de compras.
- Se implementó la vista de confirmación de compra exitosa.
- Se mejoró la presentación visual (UI/UX) del resumen del pedido y pantalla final usando Bootstrap Icons.

+++ Ecommerce/Ecommerce/Models/ShowCartViewModel.cs
@@ -15,7 +15,7 @@ namespace Ecommerce.Models
 
         //[DisplayFormat(DataFormatString = "{0:N2}")]
         [Display(Name = "Cantidad")]
-        public float Quantity => TemporalSales == null ? 0 : TemporalSales.Sum(ts => ts.Quantity);
+        public int Quantity => TemporalSales == null ? 0 : TemporalSales.Sum(ts => ts.Quantity);
 
         [DisplayFormat(DataFormatString = "{0:C2}")]
         [Display(Name = "Valor")]
